aqs_mod.py    fixed bug in which the time and time_local column were being switched    when the columns were renamed.    created a method for renaming columns for the hourly data.    broke the add_data method into two parts.    add_data now loads the data and then calls add_data2 in order to process the data further.    This facilitates testing of the code in add_data2 with a test csv file.
svobs.py
   added testing to the find method.
   changed some of the messages which are printed out to indicate
   if csv file is loaded.

svhy.py
   changed documentation.
   removed some functions that were not used.
   moved results function to its own file.

tools.py in get_info removed one of the columns dropped.

sverify.py added --test option for testing the processing of the observations.

added testaqs.csv to the data folder. This is a file created by
downloading the csv file from the EPA website and using head to remove
the first 200 lines. This creates a small file which can be used for
testing.